backend/webinars/requirements.txt
psycopg2-binary==2.9.9
boto3==1.35.0